requirements.txt
streamlit~=1.47.1
langchain-community~=0.3.27
langchain-openai~=0.3.28
python-dotenv~=1.1.1
langchain-core~=0.3.72
langchain~=0.3.27
langgraph~=0.6.3
pandas~=2.2.0
openpyxl~=3.1.2